+++ docs/change_log.md
@@ -77,6 +77,7 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 - many updates to the web UI to accomodate molecular datasets and workflows
 - add `simmate engine` commands to help with tags and different queues
 - add docs to help with simmate workers, clusters, and tagging
+- add `simmate-vasp` command for common VASP utilities
 
 **Refactors**